Touch up `Style/AlignHash` cop
This change does the following:

 - Port the documentation from configuration file to YARD.
 - Extract hash alignment checkers into its own `Mixin::HashAlignment` module.
 - Change abstract base class `ValueAlignment` into a mixin.
 - Remove some unneeded code from `Style/AlignHash`.
 - Reduce direct access to instance variables.

There is still quite a bit of room for improvement. For example, the cop uses
an instance variable as a temporary register to pass data to
`#good_alignment?`, the `TableAlignment` checker relies on methods being
called in a certain order, etc.
