+++ lib/rubocop.rb
@@ -64,6 +64,7 @@ require 'rubocop/cop/mixin/end_keyword_alignment'
 require 'rubocop/cop/mixin/first_element_line_break'
 require 'rubocop/cop/mixin/frozen_string_literal'
 require 'rubocop/cop/mixin/hash_node'
+require 'rubocop/cop/mixin/hash_alignment'
 require 'rubocop/cop/mixin/if_node'
 require 'rubocop/cop/mixin/integer_node'
 require 'rubocop/cop/mixin/on_method_def'
